Relative Size Caching Update
The system looks at the size of the user's window and changes pixel-based distances based on the actual size of the window. However, the system was not properly caching those values, so the math was being done over and over again, which is a performance drain. Now, cache calculations happen 1) when the application is launched, and 2) when the user changes the size of the window. Those values are then stored, significantly improving performance.