Font Fix
There was an issue with the AI generated font loader (and by proxy the GUI system that loads the text elements like a "text button") so I hard-coded it. You can simply change the names of the font directory and files. I'm working with the AI to see if I can solve this dynamically, but for now this works.